+++ lib/models/income.dart
@@ -15,7 +15,7 @@ class Income {
     return Income(
       id: map['id'],
       name: map['name'],
-      amount: map['amount'],
+      amount: map['amount'].toDouble(),
       date: DateTime.parse(map['date']),
     );
   }
